Use int64 for solutions

+++ 2024/Advent.fs
@@ -4,6 +4,6 @@ module Advent
 type DayFunctions<'input> = {
     Tests: Expecto.Test
     ReadInput: string -> 'input
-    PartOne: 'input -> int
-    PartTwo: 'input -> int
+    PartOne: 'input -> int64
+    PartTwo: 'input -> int64
 }
